requirements.txt
botometer==1.6.1
python-dotenv==1.0.0
secure-smtplib==0.1.1
streamlit==1.16.0
tweepy==3.10.0
